refactor move service to its respective file
